Refactor useGameLoop to follow best practices
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/gust42/idlefootball?shareId=XXXX-XXXX-XXXX-XXXX).
